Api/requirements.txt
tensorflow
fastapi
uvicorn
python-multipart
pillow
numpy
keras